+++ server/src/main/java/pt/ulisboa/tecnico/sec/candeeiros/server/model/CheckAccountIntent.java
@@ -32,7 +32,7 @@ public class CheckAccountIntent {
             majority = timestamp;
         }
 
-        return occurrences.get(majority) >= (Math.ceil(totalServers / 2));
+        return occurrences.get(majority) >= (Math.ceil((double)totalServers / 2));
     }
 
     public Bank.CheckAccountResponse getMajority() {
