+++ server/src/main/java/pt/ulisboa/tecnico/sec/candeeiros/server/model/sendAmountIntent.java
@@ -61,7 +61,7 @@ public class sendAmountIntent {
     }
 
     public boolean hasMajority(int totalServers) {
-        return occurrences.get(majority) >= (Math.ceil(totalServers/2));
+        return occurrences.get(majority) >= (Math.ceil((double)totalServers/2));
     }
 
     public Bank.SendAmountResponse.Status getMajority() {
